+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Added
+  * Added new prop to apply column highlighting style intended for timelines
+
 ## 2.34.0 - (November 16, 2021)
 
 * Changed